requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
pdfplumber==0.11.4
pillow==10.4.0
